# docs/openapi.front.yaml
openapi: 3.1.0
info:
  title: Lianchuan Storefront API
  version: 0.1.0
  summary: 前台电商网站接口契约
  description: |
    本文档定义前台商城所需的 REST API 契约，覆盖首页、分类、搜索、商品详情、购物车、结账、地址、订单、认证、会员资料、询价、收藏等能力。
    该文档是前台与服务端联调的唯一接口真源。
servers:
  - url: http://localhost:5100
    description: Local admin API (vexmotor-admin)
  - url: https://example.com
    description: Production placeholder
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Front-office JWT from /api/front/auth/login or OAuth callback
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SearchKeywordParam:
      name: keyword
      in: query
      schema:
        type: string
        minLength: 1
...
    LocaleHeader:
      name: x-vex-locale
      in: header
      description: 前台语言代码；与 site_languages 绑定，同时决定价格币种
      schema:
        type: string
        example: en
    LocaleQueryParam:
      name: locale
      in: query
      description: 可选语言覆盖（优先级低于 x-vex-locale）
      schema:
        type: string
        example: en
...
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
      required: [page, pageSize, total, totalPages]
      additionalProperties: false
    Money:
      type: object
      properties:
        currency:
          type: string
          example: USD
        amount:
          type: number
        formatted:
          type: string
      required: [currency, amount, formatted]
      additionalProperties: false
    ImageAsset:
      type: object
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        alt:
          type: string
        width:
          type: integer
        height:
          type: integer
      required: [id, url, alt]
      additionalProperties: false
    Brand:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        logo:
          $ref: '#/components/schemas/ImageAsset'
      required: [id, name, slug]
      additionalProperties: false
...
    ProductCard:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        sku:
          type: string
        shortDescription:
          type: string
        coverImage:
          $ref: '#/components/schemas/ImageAsset'
        price:
          $ref: '#/components/schemas/Money'
        compareAtPrice:
          oneOf:
            - $ref: '#/components/schemas/Money'
            - type: 'null'
        purchaseMode:
          type: string
          enum: [buy, inquiry]
        inStock:
          type: boolean
        brand:
          oneOf:
            - $ref: '#/components/schemas/Brand'
            - type: 'null'
      required: [id, name, slug, sku, purchaseMode, inStock]
      additionalProperties: false
...
    ProductListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProductCard'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
        facets:
          type: array
          items:
            type: object
            additionalProperties: true
      required: [items, meta, facets]
      additionalProperties: false
...
    CompanyProfile:
      type: object
      properties:
        company:
          type:
            - string
            - 'null'
        industry:
          type:
            - string
            - 'null'
        companyCountryCode:
          type:
            - string
            - 'null'
        companyState:
          type:
            - string
            - 'null'
        companyCity:
          type:
            - string
            - 'null'
        companyAddressLine1:
          type:
            - string
            - 'null'
        companyAddressLine2:
          type:
            - string
            - 'null'
        companyPostalCode:
          type:
            - string
            - 'null'
        website:
          type:
            - string
            - 'null'
        taxId:
          type:
            - string
            - 'null'
        companySize:
          type:
            - string
            - 'null'
      additionalProperties: false
...
    CompareProductSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        spu:
          type: string
        brand:
          type:
            - string
            - 'null'
        category:
          type:
            - string
            - 'null'
        price:
          $ref: '#/components/schemas/Money'
        stockQuantity:
          type: integer
        inStock:
          type: boolean
        purchaseMode:
          type: string
      required: [id, name, slug, spu, price, stockQuantity, inStock, purchaseMode]
      additionalProperties: false
    CompareMatrixRow:
      type: object
      properties:
        key:
          type: string
        label:
          type: string
        group:
          type: string
        values:
          type: array
          items:
            type: string
        isDifferent:
          type: boolean
      required: [key, label, group, values, isDifferent]
      additionalProperties: false
    CompareResult:
      type: object
      properties:
        locale:
          type: string
        products:
          type: array
          items:
            $ref: '#/components/schemas/CompareProductSummary'
        groups:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              rows:
                type: array
                items:
                  type: string
            required: [name, rows]
        matrix:
          type: array
          items:
            $ref: '#/components/schemas/CompareMatrixRow'
        summary:
          type: object
          properties:
            sharedFeatureKeys:
              type: array
              items:
                type: string
            uniqueByProduct:
              type: array
              items:
                type: object
                properties:
                  productId:
                    type: string
                    format: uuid
                  keys:
                    type: array
                    items:
                      type: string
                required: [productId, keys]
          required: [sharedFeatureKeys, uniqueByProduct]
      required: [locale, products, groups, matrix, summary]
      additionalProperties: false
...
paths:
...
  /api/front/search:
    get:
      tags: [Search]
      summary: 搜索商品
      parameters:
        - $ref: '#/components/parameters/SearchKeywordParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: 搜索结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
...
  /api/front/profile/company:
    get:
      tags: [Profile]
      summary: 获取当前用户公司资料
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 公司资料
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyProfile'
    patch:
      tags: [Profile]
      summary: 更新当前用户公司资料
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyProfile'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyProfile'
...
  /api/front/compare:
    post:
      tags: [Compare]
      summary: 执行商品对比矩阵（2-4 个商品，无需先加入列表）
      parameters:
        - $ref: '#/components/parameters/LocaleHeader'
        - $ref: '#/components/parameters/LocaleQueryParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                productIds:
                  type: array
                  minItems: 2
                  maxItems: 4
                  items:
                    type: string
                    format: uuid
              required: [productIds]
      responses:
        '200':
          description: 对比结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompareResult'